refactor, fetching map loader fixes, snackbar fixes
