+++ src/components/stateful/GoogleMapHandler/index.js
@@ -1,9 +1,10 @@
 import React, { Component } from 'react';
 import MapLayout from "./MapLayoutComponent";
 import PropTypes from 'prop-types';
+import {connect} from "react-redux";
 
 const mapUrl = "https://maps.googleapis.com/maps/api/js?key=AIzaSyCac4KiePeteNK02OJTgHWGtBEvMAWGL5M&v=3.exp&libraries=geometry,drawing,places"
-// const mapUrl = "https://maps.googleapis.com/maps/api/js?v=3.exp&libraries=geometry,drawing,places"
+
 const markerItems = [
     {
         position: {
@@ -23,10 +24,8 @@ const markerItems = [
 ];
 
 class GoogleMapHandler extends Component {
-
     render() {
-        const { onMarkerClick, onMarkerClusterClick, showBicyclelayer, dataSource  } = this.props;
-
+        const { onMarkerClick, onMarkerClusterClick, showBicyclelayer  } = this.props;
         return (
             <MapLayout
                 googleMapURL={mapUrl}
@@ -37,7 +36,6 @@ class GoogleMapHandler extends Component {
                 onMarkerClusterClick={onMarkerClusterClick}
                 markers={markerItems}
                 showBicyclelayer={showBicyclelayer}
-                dataSource={dataSource}
             />
         )
     }
@@ -47,11 +45,7 @@ GoogleMapHandler.propTypes = {
     containerHeight: PropTypes.number,
     onMarkerClick: PropTypes.func,
     onMarkerClusterClick: PropTypes.func,
-    showBicyclelayer: PropTypes.bool,
-    dataSource: PropTypes.shape({
-        items: PropTypes.array.isRequired,
-        isFetching: PropTypes.bool.isRequired
-    })
+    showBicyclelayer: PropTypes.bool
 };
 
-export default GoogleMapHandler;
+export default connect()(GoogleMapHandler);
