+++ src/containers/user/Dashboard/components/stateless/ConfirmDialog/index.js
