+++ src/containers/user/Dashboard/index.js
@@ -1,39 +1,41 @@
 import React, { Component } from 'react';
 import './styles.css';
-import GoogleMapHandler from "../../../../components/stateful/GoogleMapHandler";
+import GoogleMapHandler from "../../../components/stateful/GoogleMapHandler";
 import DashboardListContainer from "./components/stateless/DashboardList/index";
 import { connect } from "react-redux";
-import { actionMapDataSource, actionMapisFetching } from "../../../../actions/action-map";
+import { actionMapDataSource, actionMapisFetching } from "../../../actions/action-map";
 import axios from "axios/index";
-import ConfirmDialog from "./components/stateless/ConfirmDialog"
-// import { NavigationFullscreen } from 'material-ui';
+import ConfirmDialog from "./components/stateless/ConfirmDialog";
 
 class UserDashboardContainer extends Component {
+
+    constructor(props) {
+        super(props);
+        this.state = {
+            markerToSelect: null,
+            beginningStation: null,
+            dockingStation: null
+        };
+    }
 
     componentDidMount() {
         this.getBikepoints();
     }
 
     getBikepoints = () => {
-        this.props.actionMapisFetching(true);
+        const { actionMapisFetching, actionMapDataSource } = this.props;
+
+        actionMapisFetching(true);
         axios.get("https://tajz77isu1.execute-api.us-east-1.amazonaws.com/dev/bikepoint", {
             responseType: 'json'
         })
-            .then(response => {
-                this.props.actionMapDataSource(response.data);
-                this.props.actionMapisFetching(false);
-                this.setState({ commonName: null });
-            })
-            .catch(error => {
-                alert(error)
-            });
+        .then(response => {
+            actionMapDataSource(response.data);
+            actionMapisFetching(false);
+            this.setState({ commonName: null });
+        })
+        .catch(error =>  alert(error) );
     };
-
-    state = {
-        markerToSelect: null,
-        beginningStation: null,
-        dockingStation: null
-    }
 
     footer = (dataSource) => {
         if (this.state.markerToSelect == null) {
@@ -48,26 +50,26 @@ class UserDashboardContainer extends Component {
                 />
             )
         }
-    }
+    };
 
     _onMarkerClick = (event) => {
-        const { marker } = event
+        const { marker } = event;
         this.setState({ markerToSelect: marker })
-    }
+    };
 
     _onMarkerClusterClick = () => {
-    }
+    };
 
     _onCancelClick = () => {
         this.setState({ markerToSelect: null })
-    }
+    };
 
     _onSelectClick = () => {
         const updatedStation = this.state.beginningStation == null ?
             { beginningStation: this.state.markerToSelect } :
-            { dockingStation: this.state.markerToSelect }
+            { dockingStation: this.state.markerToSelect };
         this.setState({ ...updatedStation, markerToSelect: null })
-    }
+    };
 
     render() {
 
@@ -76,9 +78,9 @@ class UserDashboardContainer extends Component {
         return (
             <div className="dashboard-container">
                 <GoogleMapHandler
-                    dataSource={dataSource}
                     onMarkerClick={this._onMarkerClick}
-                    onMarkerClusterClick={this._onMarkerClusterClick} />
+                    onMarkerClusterClick={this._onMarkerClusterClick}
+                />
                 {this.footer(dataSource)}
             </div>
         )
@@ -94,4 +96,5 @@ const mapDispatchToProps = dispatch => ({
 const mapStateToProps = state => ({
     dataSource: state.reducerMapDatasource
 });
+
 export default connect(mapStateToProps, mapDispatchToProps)(UserDashboardContainer)
